+++ tests/integration/events/v1/test_event_type.py
@@ -59,6 +59,7 @@ class EventTypeTestCase(IntegrationTestCase):
                         "date_updated": "2020-08-13T13:28:20Z",
                         "type": "com.twilio.messaging.message.delivered",
                         "schema_id": "Messaging.MessageStatus",
+                        "public": true,
                         "description": "Messaging- delivered message",
                         "url": "https://events.twilio.com/v1/Types/com.twilio.messaging.message.delivered",
                         "links": {
@@ -70,6 +71,7 @@ class EventTypeTestCase(IntegrationTestCase):
                         "date_updated": "2020-08-13T13:28:19Z",
                         "type": "com.twilio.messaging.message.failed",
                         "schema_id": "Messaging.MessageStatus",
+                        "public": true,
                         "description": "Messaging- failed message",
                         "url": "https://events.twilio.com/v1/Types/com.twilio.messaging.message.failed",
                         "links": {
@@ -114,6 +116,7 @@ class EventTypeTestCase(IntegrationTestCase):
                 "date_updated": "2020-08-13T13:28:20Z",
                 "type": "com.twilio.messaging.message.delivered",
                 "schema_id": "Messaging.MessageStatus",
+                "public": true,
                 "description": "Messaging- delivered message",
                 "url": "https://events.twilio.com/v1/Types/com.twilio.messaging.message.delivered",
                 "links": {