+++ tests/integration/messaging/v1/service/test_us_app_to_person.py
@@ -53,7 +53,7 @@ class UsAppToPersonTestCase(IntegrationTestCase):
                 "us_app_to_person_usecase": "MARKETING",
                 "has_embedded_links": true,
                 "has_embedded_phone": false,
-                "status": "PENDING",
+                "campaign_status": "PENDING",
                 "campaign_id": "CXaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "is_externally_registered": false,
                 "rate_limits": {
@@ -128,7 +128,7 @@ class UsAppToPersonTestCase(IntegrationTestCase):
                 "us_app_to_person_usecase": "MARKETING",
                 "has_embedded_links": true,
                 "has_embedded_phone": false,
-                "status": "PENDING",
+                "campaign_status": "PENDING",
                 "campaign_id": "CXaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "is_externally_registered": false,
                 "rate_limits": {