+++ tests/integration/verify/v2/service/entity/test_challenge.py
@@ -30,17 +30,17 @@ class ChallengeTestCase(IntegrationTestCase):
             data=values,
         ))
 
-    def test_create_without_auth_payload_response(self):
+    def test_create_push_without_auth_payload_response(self):
         self.holodeck.mock(Response(
             201,
             '''
             {
-                "sid": "YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "sid": "YC03aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "service_sid": "VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "entity_sid": "YEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "identity": "ff483d1ff591898a9942916050d2ca3f",
-                "factor_sid": "YFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "factor_sid": "YF03aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "date_created": "2015-07-30T20:00:00Z",
                 "date_updated": "2015-07-30T20:00:00Z",
                 "date_responded": "2015-07-30T20:00:00Z",
@@ -61,9 +61,9 @@ class ChallengeTestCase(IntegrationTestCase):
                     "ip": "172.168.1.234"
                 },
                 "factor_type": "push",
-                "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YC03aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
-                    "notifications": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Notifications"
+                    "notifications": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YC03aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Notifications"
                 }
             }
             '''
@@ -75,17 +75,62 @@ class ChallengeTestCase(IntegrationTestCase):
 
         self.assertIsNotNone(actual)
 
-    def test_create_with_auth_payload_response(self):
+    def test_create_totp_without_auth_payload_response(self):
         self.holodeck.mock(Response(
             201,
             '''
             {
-                "sid": "YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "sid": "YC02aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "service_sid": "VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "entity_sid": "YEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "identity": "ff483d1ff591898a9942916050d2ca3f",
-                "factor_sid": "YFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "factor_sid": "YF02aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "date_created": "2015-07-30T20:00:00Z",
+                "date_updated": "2015-07-30T20:00:00Z",
+                "date_responded": "2015-07-30T20:00:00Z",
+                "expiration_date": "2015-07-30T20:00:00Z",
+                "status": "pending",
+                "responded_reason": "none",
+                "details": {
+                    "message": "Hi! Mr. John Doe, would you like to sign up?",
+                    "date": "2020-07-01T12:13:14Z",
+                    "fields": [
+                        {
+                            "label": "Action",
+                            "value": "Sign up in portal"
+                        }
+                    ]
+                },
+                "hidden_details": {
+                    "ip": "172.168.1.234"
+                },
+                "factor_type": "totp",
+                "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YC02aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "links": {
+                    "notifications": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YC02aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Notifications"
+                }
+            }
+            '''
+        ))
+
+        actual = self.client.verify.v2.services("VAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX") \
+                                      .entities("identity") \
+                                      .challenges.create(factor_sid="YFXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX")
+
+        self.assertIsNotNone(actual)
+
+    def test_create_totp_with_auth_payload_response(self):
+        self.holodeck.mock(Response(
+            201,
+            '''
+            {
+                "sid": "YC02aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "service_sid": "VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "entity_sid": "YEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "identity": "ff483d1ff591898a9942916050d2ca3f",
+                "factor_sid": "YF02aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "date_created": "2015-07-30T20:00:00Z",
                 "date_updated": "2015-07-30T20:00:00Z",
                 "date_responded": "2015-07-30T20:00:00Z",
@@ -105,10 +150,10 @@ class ChallengeTestCase(IntegrationTestCase):
                 "hidden_details": {
                     "ip": "172.168.1.234"
                 },
-                "factor_type": "push",
-                "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "factor_type": "totp",
+                "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YC02aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
-                    "notifications": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Notifications"
+                    "notifications": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YC02aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Notifications"
                 }
             }
             '''